# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Chariot Financial Solutions case management API
servers:
  - url: /api
...
paths:
...
  /cases/{id}/service-level/confirm:
    post:
      operationId: confirmCaseServiceLevel
      tags: [cases]
      description: Administrator only. The adviser confirms the service level the case runs at (scope step 5).
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Case with the confirmation recorded
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CaseDetail"}
  /cases/{id}/advice:
    get:
      operationId: getCaseAdvice
      tags: [cases]
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: The advice draft and every version sent to the client
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CaseAdviceState"}
    put:
      operationId: updateCaseAdvice
      tags: [cases]
      description: Administrator only. Saves the advice draft.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CaseAdviceInput"}
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CaseAdviceState"}
  /cases/{id}/advice/send:
    post:
      operationId: sendCaseAdvice
      tags: [cases]
      description: Administrator only. Snapshots the advice, emails it with Approve / Further discussion buttons, and waits for the client.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CaseAdviceState"}
        "409":
          description: Service level not confirmed, advice incomplete, or this service level does not send advice
  /cases/{id}/advice/extract:
    post:
      operationId: extractCaseInstruction
      tags: [cases]
      description: Read the client's own instruction (lender, product, rate, term, amount) out of a pasted email. Nothing is saved.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExtractInstructionInput"}
      responses:
        "200":
          description: Suggested fields
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExtractInstructionOutput"}
  /cases/{id}/approvals/{approvalId}/record:
    post:
      operationId: recordCaseApproval
      tags: [cases]
      description: Staff record an answer the client gave by phone or in a reply email.
      parameters:
        - {$ref: "#/components/parameters/Id"}
        - {name: approvalId, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/RecordApprovalInput"}
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ClientApproval"}
  /cases/{id}/submission-details:
    get:
      operationId: getCaseSubmissionDetails
      tags: [cases]
      description: The structured details pack that will be submitted, what is still missing, and the client's confirmation state.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetailsState"}
  /cases/{id}/submission-details/send:
    post:
      operationId: sendCaseSubmissionDetails
      tags: [cases]
      description: Snapshots the pack and asks the client to confirm it (email button + portal).
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SubmissionDetailsState"}
        "409":
          description: The pack still has missing fields
  /cases/{id}/prefill-from-previous:
    post:
      operationId: prefillCaseFromPrevious
      tags: [cases]
      description: Copies still-empty details from the client's most recent other case.
      parameters:
        - {$ref: "#/components/parameters/Id"}
      responses:
        "200":
          description: What was copied
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PrefillResult"}
        "404":
          description: The client has no previous case
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: integer}
...
  schemas:
...
    TermsAcceptance:
      type: object
      required: [acceptedAt, via, version, note, acceptedBy]
      properties:
        acceptedAt: {type: string, format: date-time}
        via: {type: string, enum: [portal, signed_upload, staff]}
        version: {type: integer, nullable: true}
        note: {type: string, nullable: true}
        acceptedBy: {type: string, nullable: true, description: Staff member who recorded it, when not accepted by the client in the portal.}
...
    ExtractInstructionInput:
      type: object
      required: [emailText]
      properties:
        emailText: {type: string, minLength: 1, maxLength: 40000}
    ExtractInstructionOutput:
      type: object
      required: [extracted, model]
      properties:
        model: {type: string, nullable: true}
        extracted:
          type: object
          required: [lenderId, lenderName, product, ratePct, termYears, loanAmount, summary]
          properties:
            lenderId: {type: integer, nullable: true}
            lenderName: {type: string, nullable: true}
            product: {type: string, nullable: true}
            ratePct: {type: number, nullable: true}
            termYears: {type: integer, nullable: true}
            loanAmount: {type: number, nullable: true}
            summary: {type: string, nullable: true}
...
    CaseSubmissionStatus:
      type: string
      enum: [active, withdrawn, declined, offered]
    CaseSubmission:
      type: object
      required: [id, caseId, lenderId, lenderName, status, isPrimary, lenderCaseNumber, dipDocument, applicationFeeConfirmed, valuationDate, valuationCompletedAt, bankDecisionRequested, notes, closedAt, closeReason, createdAt]
      properties:
        id: {type: integer}
        caseId: {type: integer}
        lenderId: {type: integer}
        lenderName: {type: string}
        status: {$ref: "#/components/schemas/CaseSubmissionStatus"}
        isPrimary: {type: boolean}
        lenderCaseNumber: {type: ["string", "null"]}
        dipDocument: {allOf: [{$ref: "#/components/schemas/CaseDipDocument"}], nullable: true}
        applicationFeeConfirmed: {type: boolean}
        valuationDate: {type: ["string", "null"], format: date-time}
        valuationCompletedAt: {type: ["string", "null"], format: date-time}
        bankDecisionRequested: {type: boolean}
        notes: {type: string}
        closedAt: {type: ["string", "null"], format: date-time}
        closeReason: {type: ["string", "null"]}
        createdAt: {type: string, format: date-time}
...
    CaseDipDocument:
      type: object
      required: [id, name]
      properties:
        id: {type: integer}
        name: {type: string}
...
    UnderwritingRound:
      type: object
      required: [round, emailText, createdAt]
      properties:
        round: {type: integer}
        emailText: {type: string}
        createdAt: {type: string, format: date-time}
    Case:
      type: object
      required: [id, reference, clientId, clientName, propertyAddress, propertyValue, matterType, serviceType, serviceLevelConfirmedAt, serviceLevelConfirmedBy, brokerFeeBasis, brokerFeeFlat, stage, stageIndex, stageDays, stageFlagged, status, loanAmount, assignedTo, updatedAt, skippedStageIndexes, procFeePct, brokerFeePct, procFee, brokerFee, revenue, archivedAt, lenderId, lenderName, caseNumber, dipDocument, valuationDate, valuationCompletedAt, valuationAmount, expectedCompletionDate, applicationFeeConfirmed, bankDecisionRequested, underwritingCleared, underwritingRounds]
      properties:
        id: {type: integer}
        reference: {type: string}
        clientId: {type: integer}
        clientName: {type: string}
        propertyId: {type: integer, nullable: true}
        propertyAddress: {type: string}
        propertyValue: {type: number}
        rent: {type: number, nullable: true}
        gdv: {type: number, nullable: true}
        matterType: {type: string}
        serviceType: {$ref: "#/components/schemas/ServiceType"}
        serviceLevelConfirmedAt: {type: string, format: date-time, nullable: true}
        serviceLevelConfirmedBy: {type: string, nullable: true}
        stage: {type: string}
        stageIndex: {type: integer}
        stageDays: {type: integer}
        stageThresholdDays: {type: integer, nullable: true}
        stageFlagged: {type: boolean}
        status: {type: string}
        loanAmount: {type: number}
        assignedTo: {type: string}
        updatedAt: {type: string, format: date-time}
        skippedStageIndexes: {type: array, items: {type: integer}}
        procFeePct: {type: number}
        brokerFeePct: {type: number}
        brokerFeeBasis: {type: string, enum: [percent, flat]}
        brokerFeeFlat: {type: number, nullable: true}
        procFee: {type: number}
        brokerFee: {type: number}
        revenue: {type: number}
        archivedAt: {type: string, format: date-time, nullable: true}
        lenderId: {type: integer, nullable: true}
        lenderName: {type: string, nullable: true}
        caseNumber: {type: string, nullable: true}
        dipDocument: {allOf: [{$ref: "#/components/schemas/CaseDipDocument"}], nullable: true}
        valuationDate: {type: string, format: date-time, nullable: true}
        valuationCompletedAt:
          description: When staff confirmed the valuation took place; null until confirmed.
          type: string
          format: date-time
          nullable: true
        valuationAmount:
          description: The lender's valuation figure, recorded when the valuation is confirmed; null until then.
          type: number
          nullable: true
        expectedCompletionDate:
          description: Day the mortgage is expected to complete. Mirrored to the calendar and a follow-up task.
          type: string
          format: date-time
          nullable: true
        applicationFeeConfirmed: {type: boolean}
        bankDecisionRequested: {type: boolean}
        underwritingCleared: {type: boolean}
        underwritingRounds: {type: array, items: {$ref: "#/components/schemas/UnderwritingRound"}}
...
    ServiceType:
      type: string
      enum: [full_advice, light_advice, execution_only]
    ClientApproval:
      type: object
      required: [id, caseId, kind, version, sentAt, expiresAt, deliveryStatus, deliveryError, response, respondedAt, respondedVia, note, snapshot, sentBy]
      properties:
        id: {type: integer}
        caseId: {type: integer}
        kind: {type: string, enum: [advice, submission_details]}
        version: {type: integer}
        sentAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        sentBy: {type: string, nullable: true}
        deliveryStatus: {type: string, enum: [pending, sent, disabled, failed]}
        deliveryError: {type: string, nullable: true}
        response: {type: string, enum: [approved, discuss], nullable: true}
        respondedAt: {type: string, format: date-time, nullable: true}
        respondedVia: {type: string, enum: [email, portal, staff], nullable: true}
        note: {type: string, nullable: true}
        snapshot:
          description: What the client was shown - an AdviceSnapshot or a SubmissionPack.
          type: object
          additionalProperties: true
    CaseAdvice:
      type: object
      required: [caseId, lenderId, lenderName, product, ratePct, termYears, monthlyPayment, arrangementFee, summary, reasoning, updatedAt, updatedBy, source, instructionEmailText]
      properties:
        source: {type: string, enum: [adviser, client_email, staff], nullable: true}
        instructionEmailText: {type: string, nullable: true}
        caseId: {type: integer}
        lenderId: {type: integer, nullable: true}
        lenderName: {type: string, nullable: true}
        product: {type: string, nullable: true}
        ratePct: {type: number, nullable: true}
        termYears: {type: integer, nullable: true}
        monthlyPayment: {type: number, nullable: true}
        arrangementFee: {type: number, nullable: true}
        summary: {type: string, nullable: true}
        reasoning: {type: string, nullable: true}
        updatedAt: {type: string, format: date-time, nullable: true}
        updatedBy: {type: string, nullable: true}
    CaseAdviceInput:
      type: object
      properties:
        source: {type: string, enum: [adviser, client_email, staff], nullable: true}
        instructionEmailText: {type: string, nullable: true}
        lenderId: {type: integer, nullable: true}
        product: {type: string, nullable: true}
        ratePct: {type: number, minimum: 0, nullable: true}
        termYears: {type: integer, minimum: 0, nullable: true}
        monthlyPayment: {type: number, minimum: 0, nullable: true}
        arrangementFee: {type: number, minimum: 0, nullable: true}
        summary: {type: string, nullable: true}
        reasoning: {type: string, nullable: true}
    CaseAdviceState:
      type: object
      required: [advice, approvals, latest, needsAdvice, mode, serviceLevelConfirmedAt, serviceLevelConfirmedBy, readyToSend, missing, instructionRecorded, termsOfBusiness]
      properties:
        mode:
          description: advice = the adviser writes and sends a recommendation; instruction = the client's own instruction (execution only), confirmed with the submission details.
          type: string
          enum: [advice, instruction]
        instructionRecorded: {type: boolean}
        termsOfBusiness:
          allOf: [{$ref: "#/components/schemas/TermsAcceptance"}]
          nullable: true
        advice: {$ref: "#/components/schemas/CaseAdvice"}
        approvals:
          type: array
          items: {$ref: "#/components/schemas/ClientApproval"}
        latest:
          allOf: [{$ref: "#/components/schemas/ClientApproval"}]
          nullable: true
        needsAdvice: {type: boolean, description: Whether this service level sends written advice.}
        serviceLevelConfirmedAt: {type: string, format: date-time, nullable: true}
        serviceLevelConfirmedBy: {type: string, nullable: true}
        readyToSend: {type: boolean}
        missing:
          type: array
          items: {type: string}
          description: What still stops the advice being sent.
    RecordApprovalInput:
      type: object
      required: [response]
      properties:
        response: {type: string, enum: [approved, discuss]}
        note: {type: string}
...
    SubmissionField:
      type: object
      required: [key, label, value, source, missing]
      properties:
        key: {type: string}
        label: {type: string}
        value: {type: string, nullable: true}
        source: {type: string, enum: [client, property, case, advice]}
        missing: {type: boolean}
    SubmissionSection:
      type: object
      required: [key, title, fields]
      properties:
        key: {type: string}
        title: {type: string}
        fields:
          type: array
          items: {$ref: "#/components/schemas/SubmissionField"}
    SubmissionPack:
      type: object
      required: [sections, missing]
      properties:
        sections:
          type: array
          items: {$ref: "#/components/schemas/SubmissionSection"}
        missing:
          type: array
          items: {type: string}
          description: Labels of the required fields that are still empty.
    SubmissionDetailsState:
      type: object
      required: [pack, approvals, latest, confirmed, previousCase]
      properties:
        pack: {$ref: "#/components/schemas/SubmissionPack"}
        approvals:
          type: array
          items: {$ref: "#/components/schemas/ClientApproval"}
        latest:
          allOf: [{$ref: "#/components/schemas/ClientApproval"}]
          nullable: true
        confirmed: {type: boolean}
        previousCase:
          type: object
          nullable: true
          required: [id, reference, stage, updatedAt]
          properties:
            id: {type: integer}
            reference: {type: string}
            stage: {type: string}
            updatedAt: {type: string, format: date-time}
    PrefillResult:
      type: object
      required: [copied, fromReference]
      properties:
        copied:
          type: array
          items: {type: string}
        fromReference: {type: string}
...
    CaseDetail:
      allOf:
        - {$ref: "#/components/schemas/Case"}
        - type: object
          required: [stages, requirements, tasks, messages, draftNotes, submissions]
          properties:
            submissions:
              description: Every lender this case has been submitted to, primary first. The case's own lender fields mirror the primary submission.
              type: array
              items: {$ref: "#/components/schemas/CaseSubmission"}
            stages:
              type: array
              items: {type: string}
            requirements:
              type: array
              items: {$ref: "#/components/schemas/Requirement"}
            tasks:
              type: array
              items: {$ref: "#/components/schemas/Task"}
            messages:
              type: array
              items: {$ref: "#/components/schemas/Message"}
            draftNotes: {type: string}
...
    Requirement:
      type: object
      required: [id, stageIndex, label, complete, required, round]
      properties:
        id: {type: integer}
        stageIndex: {type: integer}
        label: {type: string}
        complete: {type: boolean}
        required: {type: boolean}
        round: {type: integer}
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, normal, high, urgent]
    Task:
      type: object
      required: [id, title, caseId, clientId, caseReference, clientName, assignee, assignedUserId, status, priority, dueDate, notes, createdAt, completedAt, createdByUserId, createdByName, checklistTotal, checklistDone, checklistNext, commentCount, kind, propertyId]
      properties:
        id: {type: integer}
        title: {type: string}
        caseId: {type: ["integer", "null"]}
        clientId:
          description: Client the task concerns (set for onboarding/property tasks; derived from the case otherwise).
          type: ["integer", "null"]
        caseReference: {type: ["string", "null"]}
        clientName: {type: ["string", "null"]}
        assignee: {type: string}
        assignedUserId: {type: integer}
        status: {$ref: "#/components/schemas/TaskStatus"}
        priority: {$ref: "#/components/schemas/TaskPriority"}
        dueDate: {type: string, format: date}
        notes: {type: string}
        createdAt: {type: string, format: date-time}
        completedAt: {type: ["string", "null"], format: date-time}
        createdByUserId: {type: ["integer", "null"]}
        createdByName: {type: ["string", "null"]}
        checklistTotal: {type: integer}
        checklistDone: {type: integer}
        checklistNext:
          description: Title of the first unfinished checklist step, i.e. what the task is up to; null when there is none.
          type: ["string", "null"]
        commentCount: {type: integer}
        kind:
          description: Why the task exists (client_onboarding, property_review, case_submission, stage_handoff, submission_step, property_import); null for manual tasks.
          type: ["string", "null"]
        propertyId: {type: ["integer", "null"]}
...
    Message:
      type: object
      required: [id, caseId, caseReference, sender, senderRole, body, createdAt]
      properties:
        id: {type: integer}
        caseId: {type: integer}
        caseReference: {type: string}
        senderUserId: {type: ["integer", "null"]}
        sender: {type: string}
        senderRole: {type: string}
        body: {type: string}
        replyTo:
          type: ["object", "null"]
          allOf: [{$ref: "#/components/schemas/MessageReplyPreview"}]
        reactions: {type: array, items: {$ref: "#/components/schemas/MessageReaction"}}
        attachments: {type: array, items: {$ref: "#/components/schemas/MessageAttachment"}}
        createdAt: {type: string, format: date-time}
    MessageReplyPreview:
      type: object
      required: [id, sender, body]
      properties:
        id: {type: integer}
        sender: {type: string}
        body: {type: string}
    MessageReaction:
      type: object
      required: [emoji, count, users, reacted]
      properties:
        emoji: {type: string}
        count: {type: integer}
        users: {type: array, items: {type: string}, description: Display names of users who reacted}
        reacted: {type: boolean, description: Whether the current user added this reaction}
...
    MessageAttachment:
      type: object
      required: [id, name, contentType, byteSize]
      properties:
        id: {type: integer}
        name: {type: string}
        contentType: {type: string}
        byteSize: {type: integer}
...